Создает файлы для модулей
